+++ test/smokes/misspell/option_target/dir2/file.rb
@@ -1,0 +1,3 @@
+def infomation
+  puts 1
+end