+++ test/smokes/misspell/option_target/file2.rb
@@ -1,0 +1,3 @@
+def information
+  puts 1
+end